Fix database connection "keep alive".
This should fix the eval crashing when it is unable to connect to the database server.

We already had some keep alive system implemented, but when an attempt to reconnect failed the first time the db_connect function would log_error and the eval would stop.

Reviewed: http://reviewboard.infoarena.ro/r/96/



git-svn-id: http://hackers.devnet.ro/dav-svn/trunk@1059 ddd5704e-cc05-0410-ac10-d31acb56fd8e