Fix submitting of new tasks added to waiting rounds
There was a problem when a task owner tried to submit solutions to his task after it was added to a "waiting" round. This should fix it and still allow him to submit to a null round if there are no other options available.

Review URL: http://reviewboard.infoarena.ro/r/97/



git-svn-id: http://hackers.devnet.ro/dav-svn/trunk@1061 ddd5704e-cc05-0410-ac10-d31acb56fd8e