Fix diff between revisions sometimes erroring
The problem was caused by fread not reading the full data from pipes. It apparently reads things in chunks as their are outputed to the pipe and since we close the pipe after a single read, iconv errors and says "broken pipe".
Replaced fread with stream_get_contents as is exemplified in the php manual and everything works ok.

Review URL: http://reviewboard.infoarena.ro/r/98/



git-svn-id: http://hackers.devnet.ro/dav-svn/trunk@1062 ddd5704e-cc05-0410-ac10-d31acb56fd8e