Fix "undefined" showing up as a round when submitting.
Ran JSLint on the submit.js script and fixed what was suggested there.
We should probably do this on all js files.

The problem was caused by the for (var key in rounds) loop which
iterated through properties from the "rounds" object's prototype.
Was fixed by adding a if (rounds.hasOwnProperty(key)).

git-svn-id: http://hackers.devnet.ro/dav-svn/trunk@1108 ddd5704e-cc05-0410-ac10-d31acb56fd8e