Implemented new round type - penalty-round
Users receive scores per tasks based on the number of submissions and
the time elapsed from the beginning of the contest. The penalty for
multiple submissions, the time period after which one percentage is
deducted from a score, as well as a lower bound on a user's score
for a correctly solved problem are configurable parameters.

REVIEW URL: http://reviewboard.infoarena.ro/r/193/

git-svn-id: http://hackers.devnet.ro/dav-svn/trunk@1168 ddd5704e-cc05-0410-ac10-d31acb56fd8e